Move runCliGenerate() to test/utils
